requirements.txt
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
certifi==2024.12.14
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
kafka-python==2.0.2
ollama==0.4.6
psycopg2 @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_2d7wow2d82/croot/psycopg2_1731699769170/work
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
redis==5.2.1
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
